+++ helpers/sanitizeMentions.js
@@ -1,0 +1,3 @@
+module.exports = function(msg){
+  return msg.replace(/@everyone|@here|<@[!&]?\d+>/g, '');
+}